# backend/src/main/java/com/intelligence/assistant/controller/AuthController.java
package com.intelligence.assistant.controller;

import com.intelligence.assistant.dto.AuthRequest;
import com.intelligence.assistant.dto.AuthResponse;
import com.intelligence.assistant.dto.RegisterRequest;
import com.intelligence.assistant.dto.UserResponse;
import com.intelligence.assistant.model.User;
import com.intelligence.assistant.service.AuthService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
@Slf4j
@CrossOrigin
public class AuthController {
    
    private final AuthService authService;
    
    @PostMapping("/register")
    public ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest request) {
        log.info("Registration request for username: {}", request.getUsername());
        AuthResponse response = authService.register(request);
        return ResponseEntity.ok(response);
    }
    
    @PostMapping("/login")
    public ResponseEntity<AuthResponse> login(@Valid @RequestBody AuthRequest request) {
        log.info("Login request for username: {}", request.getUsername());
        AuthResponse response = authService.login(request);
        return ResponseEntity.ok(response);
    }
    
    @GetMapping("/me")
    public ResponseEntity<UserResponse> getCurrentUser(@AuthenticationPrincipal User user) {
        log.info("Get current user request for: {}", user.getUsername());
        UserResponse response = authService.getCurrentUser(user.getUsername());
        return ResponseEntity.ok(response);
    }
}