# backend/src/main/java/com/intelligence/assistant/dto/AuthRequest.java
package com.intelligence.assistant.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class AuthRequest {
    
    @NotBlank(message = "Username is required")
    private String username;
    
    @NotBlank(message = "Password is required")
    private String password;
}