# backend/src/main/java/com/intelligence/assistant/dto/AuthResponse.java
package com.intelligence.assistant.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {
    
    private String accessToken;
    
    private String tokenType = "Bearer";
    
    private UserResponse user;
}