# backend/src/main/java/com/intelligence/assistant/dto/UserResponse.java
package com.intelligence.assistant.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UserResponse {
    
    private Long id;
    private String username;
    private String email;
    private Boolean isActive;
    private Boolean isAdmin;
    private LocalDateTime createdAt;
}